frontend: rework switch ZIR
Moved to a more linear layout which lends itself well to exposing an iterator.
Consumers of this iterator now just have to keep track of an index into a
homogenous sequence of bodies.
Special prongs (namely `else` and `_`) are part of the iterator to allow for
a maximum amount of code sharing with regular prongs.

The new ZIR layout also enables giving switch prong items result locations
by storing the bodies of all items inside the switch encoding itself.

This commit also re-implements `AstGen` and `print_zir` for switch based on
the new layout.